refactor(prompt): move task prompt types
