+++ tests/client_prompt/test_a2a_t_task_prompt_renderer.py
@@ -105,7 +105,7 @@ class A2ATTaskPromptRendererTest(unittest.TestCase):
         )
 
     def test_render_raises_when_template_references_unknown_slot(self) -> None:
-        from a2a_t.prompt.task_rendering import TaskPromptRenderError
+        from a2a_t.prompt.task_rendering.errors import TaskPromptRenderError
         from a2a_t.prompt.task_rendering import TaskPromptRenderer
 
         renderer = TaskPromptRenderer()
@@ -120,7 +120,7 @@ class A2ATTaskPromptRendererTest(unittest.TestCase):
             )
 
     def test_render_raises_when_template_is_invalid(self) -> None:
-        from a2a_t.prompt.task_rendering import TaskPromptRenderError
+        from a2a_t.prompt.task_rendering.errors import TaskPromptRenderError
         from a2a_t.prompt.task_rendering import TaskPromptRenderer
 
         renderer = TaskPromptRenderer()
